Changed send-package customer ID to user ID, merged addresses into address model
