+++ app/Models/Package.php
@@ -15,7 +15,7 @@ class Package extends Model {
     protected $primaryKey = 'id'; // Custom primary key
     protected $fillable = [
         'reference',
-        'customer_id',
+        'user_id',
         'origin_location_id',
         'current_location_id',
         'destination_location_id',
@@ -56,7 +56,7 @@ class Package extends Model {
 
     public function address()
     {
-        return $this->belongsTo(Addresses::class, 'addresses_id');
+        return $this->belongsTo(Address::class, 'addresses_id');
     }
 
     public function originLocation()
